[#VGj71HcN] Create FormBuilder component
It should have a select field that has a dropdown of all the form
elements type (input, textarea etc), it should have a state to handle
what form element to render based on the selection of the dropdown.

It should have a field for the question of the form element. And the
data should be passed up to the parent where the FormBuilder component
would be used.

https://trello.com/c/VGj71HcN